Pre: Expose `localProperties` on `CargoExtension`.
This avoids reading `local.properties` twice, and makes it clear
_when_ `local.properties` will be processed.  I expect to also process
`Cargo.toml` files (to extract `lib.name`), so knowing when this
happens will be helpful.
